Added test option to input page + retrieve all update data in result
Changes:
- Added radio selection for the tests (1, 2, 3, or final).
- Refactored client_result so that the students are parsed before it checkes the mode (doesn't matter which mode is picked this needs to be done first anyway).
- Added more code to UPDATE so that the class, test, and grade are pulled from the $_POST.

ISSUES:
- When you use SELECT mode it doesn't send the students names/id's. Works fine for UPDATE though...

+++ client_result.php
@@ -1,53 +1,80 @@
 <?php
-  $response = "'" .$_POST['mode']. "'";
-  echo "Atempting to run " .$response. " on the database...<br>";
-
-  if ($_POST['mode'] == "SELECT") {
+  // check if operator SELECT or UPDATE was chosen
+  if (key_exists('mode', $_POST) == false) {
+    echo "No operation selected, please try again with either SELECT or UPDATE selected."; //print message that no operator was selected
+  }
+  else {
+    $response = "'" .$_POST['mode']. "'";
+    echo "Atempting to run " .$response. " on the database...<br>";
+  
+    // grab list of students to process (working for UPDATE but not SELECT...)
     $students= $_POST['students'];
-    echo "<br>SELECT selected.<br>";
     echo "Students entered= " .$students. "<br><br>";
-
-    // No student names/id's
-    if ($students == '') {
+  
+    // check that student string isn't empty
+    $trimmed_students = trim($students);
+    if ($trimmed_students == '') {
       echo "No students entered, please try again with at least 1 student name/id.";
     }
     else {
       echo "Splitting students string...<br>";
-      $student_array= explode(",", $students);
+      // seperate the student names/id's into an array
+      $student_array= explode(",", $trimmed_students);
+
+      // trim any whitespace in the students name
       foreach ($student_array as $value) {
         $trimmed = trim($value);
         
-        // Ensure that there is a string to check
+        // ensure that there is a string to check
         if($trimmed != '') {
           echo "Student= " .$trimmed. ": ";
-          //determine if the student was referenced by id or name
-          if(ctype_digit($trimmed[0])) {
-            echo "Id Entered<br>"; //not detecting properly...
+        }
+      }
+      
+      // if operator SELECT was picked
+      if ($_POST['mode'] == "SELECT") {
+        echo "<br>SELECT selected.<br>";
+    
+        //loop through names
+        foreach ($student_array as $value) {
+          // id was entered
+          if(ctype_digit($value[0])) {
+            echo "Id Entered<br>";
             // call backend function to get by id.
-          } 
+          }
+          // name was entered
           else {
             echo "Name Entered<br>";
             // call backend function to get by name.
           }
         }
       }
+      // if operator UPDATE was picked
+      else if ($_POST['mode'] == "UPDATE") {
+        echo "<br>UPDATE selected.<br>";
+
+        $class= $_POST['students_class'];
+        echo "Class= " .$class. "<br><br>";
+
+        $test= $_POST['student_test'];
+        echo "Test to update is= " .$test. "<br><br>";
+
+        $grade= $_POST['students_grade'];
+        echo "Grade will be updated= " .$grade. "<br><br>";
+    
+        // id was entered
+        if(ctype_digit($student_array[0][0])) {
+          echo "Id Entered<br>";
+          // call backend function to update the selected test
+        }
+        // named was entered
+        else {
+          echo "Appologies, but you must use the studen't Id to update a grade. Please try again with the students Id.";
+        }
+      }
     }
   }
-  else if ($_POST['mode'] == "UPDATE") {
-    $students= $_POST['students'];
-    echo "<br>UPDATE selected.<br>";
-
-    // No student names/id's
-    if ($students == '') {
-        echo "No students entered. Please enter a student name or ID for changing.";
-    }
-    else {
-    }
-  }
-
-  # Call database with query...
-  $sql_query = $response. " other parts of the query";
-
+  
   #should pass back previous input (if I get time to figure out)
 ?>
